move more temporarily broken models into broken_models subfolder
