+++ guidance/models/broken_models/_lite_llm.py
@@ -1,7 +1,7 @@
 import tiktoken
 
-from ._engine._engine import Chat, Instruct
-from ._grammarless import GrammarlessTokenizer, GrammarlessEngine, Grammarless
+from .._engine._engine import Chat, Instruct
+from .._grammarless import GrammarlessTokenizer, GrammarlessEngine, Grammarless
 
 
 class LiteLLMEngine(GrammarlessEngine):
