+++ guidance/models/broken_models/_togetherai.py
@@ -1,10 +1,10 @@
 import os
-from ._engine._engine import Chat, Instruct
-from ._openai import (
+from .._engine._engine import Chat, Instruct
+from .._openai import (
     OpenAI,
     OpenAIEngine,
 )
-from .transformers import TransformersTokenizer
+from ..transformers import TransformersTokenizer
 
 
 class TogetherAI(OpenAI):
